+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * fix: resolve relative paths in config object
 * fix: allow docker and docker-compose run if WorkDirOuter or IdentityDirOuter does not exist
+* make github release happen on CI
 
 ### 0.2.0 (2019-Jan-25)